Fix method arg type definition mismatch
- These methods take NSInteger types and not plain
  integer types.
- This fixes a warning.

+++ PlayPanel.m
@@ -30,7 +30,7 @@
 @implementation PlayPanel
 
 - (id)initWithContentRect:(NSRect)contentRect
-                styleMask:(unsigned int)styleMask
+                styleMask:(NSUInteger)styleMask
                   backing:(NSBackingStoreType)bufferingType
                     defer:(BOOL)deferCreation
 {
